feat: 서버 컴포넌트에서 쿠키의 액세스 토큰을 확인하고, 로그인 상태에 따라 다르게 렌더링
